+++ src/__init__.py
@@ -1,5 +1,5 @@
 from .api import start_server
-from .video import track
+from .vision import track
 from .utils import list_cameras, PersonCounter
 
 __all__ = ["track", "start_server", "list_cameras", "PersonCounter"]